creates the user's firestore doc and updates the displayName in the user's auth record, fixed bug causing error toast to launch when the app is first opened/installed, added a Loading spinner
